@@ -160,8 +224,8 @@
     "jspm": "^0.17.0-beta.29",
     "node-static": "^0.7.7",
     "tsconfig-glob": "^0.2.1",
-    "typedoc": "^0.4.5",
-    "typescript": "^2.0.0",
+    "typedoc": "^0.5.5",
+    "typescript": "^2.1.4",
     "typescript-formatter": "^2.3.0",
     "uglify-js": "^2.7.3"
   }
