@@ -1,5 +1,12 @@
 import { Matrix4, JulianDate, Cartesian3, Cartographic, Quaternion, CesiumMath } from './cesium/cesium-imports'
 
+export const AVERAGE_HUMAN_HEIGHT = 1.77;
+
+export const EYE_ENTITY_ID = 'ar.eye'
+export const PHYSICAL_EYE_ENTITY_ID = 'ar.physical-eye' 
+export const STAGE_ENTITY_ID = 'ar.stage'
+export const PHYSICAL_STAGE_ENTITY_ID = 'ar.physical-stage' 
+
 /**
  * Describes the role of an [[ArgonSystem]]
  */
@@ -62,10 +69,10 @@ export { Role }
 /**
  * Configuration options for an [[ArgonSystem]] 
  */
-export interface Configuration {
-    version?: number,
-    uri?: string,
-    title?: string,
+export abstract class Configuration {
+    version?: number[];
+    uri?: string;
+    title?: string;
     role?: Role;
     protocols?: string[];
     userData?: any;
@@ -73,7 +80,6 @@ export interface Configuration {
     defaultUI?: {
         disable?: boolean
     };
-    'needsGeopose'?: boolean;
     'supportsCustomProtocols'?: boolean;
 }
 
@@ -81,16 +87,13 @@ export interface Configuration {
  * Viewport values are expressed using a right-handed coordinate system with the origin
  * at the bottom left corner.
  */
-export interface Viewport {
-    x: number,
-    y: number,
-    width: number,
-    height: number
-}
+export class Viewport {
+    x: number;
+    y: number;
+    width: number;
+    height: number;
 
-export namespace Viewport {
-    export function clone(viewport?:Viewport, result:Viewport=<any>{}) {
-        if (!viewport) return undefined;
+    static clone(viewport:Viewport, result:Viewport=<any>{}) {
         result.x = viewport.x;
         result.y = viewport.y;
         result.width = viewport.width;
@@ -98,7 +101,30 @@ export namespace Viewport {
         return result;
     }
 
-    export function equals(viewportA?:Viewport, viewportB?:Viewport) {
+    static equals(viewportA?:Viewport, viewportB?:Viewport) {
+        return viewportA && viewportB && 
+        CesiumMath.equalsEpsilon(viewportA.x, viewportB.x, CesiumMath.EPSILON7) &&
+        CesiumMath.equalsEpsilon(viewportA.y, viewportB.y, CesiumMath.EPSILON7) &&
+        CesiumMath.equalsEpsilon(viewportA.width, viewportB.width, CesiumMath.EPSILON7) &&
+        CesiumMath.equalsEpsilon(viewportA.height, viewportB.height, CesiumMath.EPSILON7);
+    }
+}
+
+export class NormalizedViewport {
+    x: number;
+    y: number;
+    width: number;
+    height: number;
+
+    static clone(viewport:NormalizedViewport, result:NormalizedViewport=<any>{}) {
+        result.x = viewport.x;
+        result.y = viewport.y;
+        result.width = viewport.width;
+        result.height = viewport.height;
+        return result;
+    }
+
+    static equals(viewportA?:NormalizedViewport, viewportB?:NormalizedViewport) {
         return viewportA && viewportB && 
         CesiumMath.equalsEpsilon(viewportA.x, viewportB.x, CesiumMath.EPSILON7) &&
         CesiumMath.equalsEpsilon(viewportA.y, viewportB.y, CesiumMath.EPSILON7) &&
@@ -135,24 +161,30 @@ export enum SubviewType {
 /**
  * A serialized notation for the position, orientation, and referenceFrame of an entity.
  */
-export interface SerializedEntityPose {
+export interface SerializedEntityState {
     p: Cartesian3, // position
     o: Quaternion, // orientation
     r: number | string, // reference frame
+    meta?: any // meta data
+}
+
+export namespace SerializedEntityState {
+    export function clone(state?:SerializedEntityState, result?:SerializedEntityState) {
+        if (!state) return undefined;
+        result = result || <SerializedEntityState><any>{};
+        result.p = Cartesian3.clone(state.p, result.p);
+        result.o = Quaternion.clone(state.o, result.o);
+        result.r = state.r;
+        result.meta = state.meta;
+        return result;
+    }
 }
 
 /**
  * A map of entity ids and their associated poses.
  */
-export interface SerializedEntityPoseMap {
-    [id: string]: SerializedEntityPose | undefined
-}
-
-export interface SerializedFrustum {
-    xOffset?: number,
-    yOffset?: number,
-    fov: number,
-    aspectRatio?: number
+export interface SerializedEntityStateMap {
+    [id: string]: SerializedEntityState | undefined
 }
 
 /**
@@ -167,91 +199,131 @@ export interface SerializedSubview {
     /**
      * The viewport for this subview (relative to the primary viewport)
      */
-    viewport: Viewport
+    viewport: NormalizedViewport
     /**
      * The pose for this subview (relative to the primary pose)
      */
-    pose?: SerializedEntityPose, // if undefined, identity is assumed
+    pose?: SerializedEntityState, // if undefined, identity is assumed
 }
 
+
 /**
- * The device state informs a [[REALITY_VIEWER]] about the current physical
- * configuration of the system, so that it may present a view that is compatible. 
- * The [[REALITY_VIEWER]] should consider the viewport and frustum values to be 
- * suggestions, unless their respective `strict` properties are true. The various 
- * poses and related accuracies represent the physical configuration of the system
- * at the (current) specified time. The [[REALITY_VIEWER]] may reuse the `time` 
- * and `pose` when presenting a view, though this is not mandatory 
- * (or necessarily expected).
+ * The serialized rendering parameters for a particular subview
  */
-export interface DeviceState {
-    cartographicPosition?: Cartographic;
-    geolocationAccuracy: number|undefined;
-    altitudeAccuracy: number|undefined;
+export interface ReadonlySerializedSubview {
+    readonly type: SubviewType,
+    /**
+     * The projection matrix for this subview
+     */
+    readonly projectionMatrix: Readonly<Matrix4>,
+    /**
+     * The viewport for this subview (relative to the primary viewport)
+     */
+    readonly viewport: Readonly<Viewport>
+    /**
+     * The pose for this subview (relative to the primary pose)
+     */
+    readonly pose?: Readonly<SerializedEntityState>, // if undefined, identity is assumed
+}
+
+export namespace SerializedSubview {
+    export function clone(subview:SerializedSubview, result?:SerializedSubview) {
+        result = result || <SerializedSubview>{};
+        result.type = subview.type;
+        result.projectionMatrix = Matrix4.clone(subview.projectionMatrix, result.projectionMatrix);
+        result.viewport = NormalizedViewport.clone(subview.viewport, result.viewport)!;
+        result.pose = subview.pose ? SerializedEntityState.clone(subview.pose, result.pose) : undefined;
+        return result;
+    }
+}
+
+export interface SerializedDeviceState {
+    currentFov: number;
+    eyeCartographicPosition: Cartographic|undefined;
+    eyeHorizontalAccuracy: number|undefined;
+    eyeVerticalAccuracy: number|undefined;
     viewport: Viewport;
     subviews: SerializedSubview[];
-    strict?: boolean;
+    strictSubviews: boolean;
+    isPresentingHMD: boolean;
 }
 
-/**
- * The view state is provided by a [[REALITY_VIEWER]], and describes how a 
- * [[REALITY_AUGMENTER]] should render the current frame.
- */
-export interface ViewState {
-    /**
-     * The absolute time when this view state was created
-     */
-    time: JulianDate,
+// export interface PhysicalViewState {
+//     time: JulianDate,
+//     stagePose: SerializedEntityPose|undefined,
+//     stageHorizontalAccuracy: number|undefined,
+//     stageVerticalAccuracy: number|undefined,
+//     eyePose: SerializedEntityPose|undefined,
+//     eyeCompassAccuracy: number|undefined,
+//     subviews: SerializedSubviewList,
+//     strict:boolean;
+// }
 
-    /**
-     * The viewing pose. May or may not match the 
-     * real-world (physical) viewing pose.
-     */
-    pose: SerializedEntityPose|undefined,
+// export interface ViewState {
+//     /**
+//      * The viewing pose.
+//      */
+//     pose: SerializedEntityState|undefined,
 
-    /** 
-     * The radius (in meters) of latitudinal and longitudinal uncertainty,
-     * in relation to the FIXED reference frame. Value is greater 
-     * than 0 or undefined. 
-     */
-    geolocationAccuracy: number|undefined,
+//     /**
+//      * The viewport to render into. In a DOM environment, 
+//      * the bottom left corner of the document element (document.documentElement) 
+//      * is the origin. 
+//      */
+//     viewport: Viewport,
 
-    /**
-     * The accuracy of the altitude in meters. Value is greater 
-     * than 0 or undefined. 
-     */
-    altitudeAccuracy: number|undefined,
+//     /**
+//      * The list of subviews to render.
+//      */
+//     subviews:SerializedSubviewList,
 
-    /**
-     * The accuracy of the compass in degrees. Value is greater 
-     * than 0 or undefined. 
-     */
-    compassAccuracy: number|undefined,
+//     /**
+//      * The current field of view (of each subview)
+//      */
+//     fovs: number[]
+// }
 
-    /**
-     * The source reality viewer. 
-     */
-    reality?: string,
-
-    /**
-     * The primary viewport to render into. In a DOM environment, 
-     * the bottom left corner of the document element (document.documentElement) 
-     * should be considered the origin. 
-     */
-    viewport: Viewport,
-
-    /**
-     * The list of subviews to render.
-     */
-    subviews: SerializedSubview[]
+export class SerializedSubviewList extends Array<SerializedSubview> {
+    constructor() {
+        super();
+    }
+    static clone(subviews:SerializedSubviewList, result?:SerializedSubviewList) {
+        result = result || new SerializedSubviewList;
+        result.length = subviews.length;
+        for (let i=0; i < subviews.length; i++) {
+            const s = subviews[i];
+            result![i] = SerializedSubview.clone(s, result![i]);
+        }
+        return result;
+    }
 }
+
+
+// export namespace ViewState {
+//     export function clone(viewState:ViewState, result?:ViewState) {
+//         result = result || <ViewState>{};
+//         result.time = JulianDate.clone(viewState.time, result.time);
+//         result.stagePose = viewState.stagePose ? SerializedEntityState.clone(viewState.stagePose, result.stagePose) : undefined;
+//         result.pose = viewState.pose ? SerializedEntityState.clone(viewState.pose, result.pose) : undefined;
+//         result.viewport = Viewport.clone(viewState.viewport, result.viewport);
+//         result.subviews = result.subviews || [];
+//         result.fovs = result.fovs || [];
+//         viewState.subviews.forEach((s, i)=>{
+//             result!.subviews[i] = SerializedSubview.clone(s, result!.subviews[i]);
+//         });
+//         viewState.fovs.forEach((f, i)=>{
+//             result!.fovs[i] = f
+//         });
+//         return result;
+//     }
+// }
 
 /**
  * Describes the pose of a reality view and how it is able to render
  */
 export interface DeprecatedEyeParameters {
     viewport?: Viewport; // default: maximum
-    pose?: SerializedEntityPose;
+    pose?: SerializedEntityState;
     stereoMultiplier?: number; // default: 1
     fov?: number; // default: defer to manager
     aspect?: number; // default: matches viewport 
@@ -265,17 +337,20 @@ export interface DeprecatedEyeParameters {
 export interface DeprecatedPartialFrameState {
     index: number,
     time: { dayNumber: number, secondsOfDay: number }, // JulianDate,
-    view?: ViewState,
+    view?: any,
     eye?: DeprecatedEyeParameters,
-    entities?: SerializedEntityPoseMap
+    entities?: SerializedEntityStateMap
 }
 
 /**
  * Describes a complete frame state which is sent to child sessions
  */
 export interface FrameState {
-    view: ViewState,
+    time: JulianDate,
+    viewport: Viewport,
+    subviews: SerializedSubviewList,
+    reality?: string,
     index?: number,
-    entities?: SerializedEntityPoseMap,
+    entities: SerializedEntityStateMap,
     sendTime?: { dayNumber: number, secondsOfDay: number }, // the time this state was sent
 }
