@@ -1,4 +1,4 @@
-import { inject } from 'aurelia-dependency-injection'
+import { autoinject } from 'aurelia-dependency-injection'
 import {
     Entity,
     EntityCollection,
@@ -6,15 +6,28 @@ import {
     ConstantProperty,
     Transforms,
     Cartesian3,
-    Cartographic,
     Quaternion,
+    Matrix3,
+    Matrix4,
     JulianDate,
     ReferenceFrame,
+    PerspectiveFrustum,
     defined
 } from './cesium/cesium-imports'
-import { SerializedEntityPose, SerializedEntityPoseMap, FrameState, Role, ViewState } from './common'
+import { 
+    AVERAGE_HUMAN_HEIGHT,
+    SerializedEntityState, 
+    SerializedSubviewList, 
+    SerializedEntityStateMap,
+    SubviewType,
+    FrameState,
+    Viewport, 
+    Role,
+    EYE_ENTITY_ID, PHYSICAL_EYE_ENTITY_ID,
+    STAGE_ENTITY_ID, PHYSICAL_STAGE_ENTITY_ID
+} from './common'
 import { SessionService, SessionPort } from './session'
-import { Event, getSerializedEntityPose, getEntityPositionInReferenceFrame, getEntityOrientationInReferenceFrame } from './utils'
+import { Event, getSerializedEntityState, getEntityPositionInReferenceFrame, getEntityOrientationInReferenceFrame, deprecated } from './utils'
 
 /**
  * Describes the current pose of an entity relative to a particular reference frame
@@ -38,9 +51,13 @@ export enum PoseStatus {
     LOST = 4
 }
 
-const scratchCartesian3 = new Cartesian3(0, 0);
+const scratchCartesian = new Cartesian3(0, 0);
+const scratchCartesian2 = new Cartesian3(0, 0);
 const scratchQuaternion = new Quaternion(0, 0);
-const scratchOriginCartesian3 = new Cartesian3(0, 0);
+const scratchOriginCartesian = new Cartesian3(0, 0);
+const scratchFrustum = new PerspectiveFrustum();
+const scratchMatrix3 = new Matrix3;
+const scratchMatrix4 = new Matrix4;
 
 /**
  * Provides a means of querying the current state of reality. 
@@ -59,19 +76,71 @@ const scratchOriginCartesian3 = new Cartesian3(0, 0);
  *  * `ar.context.update` - Indicates to this context that the session wants
  *    to be focused on.
  */
-@inject(SessionService)
+@autoinject()
 export class ContextService {
+
+    constructor(
+        private sessionService: SessionService
+    ) {
+        this.sessionService.manager.on['ar.context.update'] = (state: FrameState) => {
+            // backwards-compat
+            if (typeof state.reality !== 'string') {
+                state.reality = state.reality && state.reality['uri'];
+            }
+            if (!state.viewport && state['view'] && state['view'].viewport) {
+                state.viewport = state['view'].viewport;
+            }
+            if (!state.subviews && state['view'] && state['view'].subviews) {
+                state.subviews = state['view'].subviews;
+                scratchFrustum.near = 0.01;
+                scratchFrustum.far = 10000000;
+                for (const s of state.subviews) {
+                    const frustum = s['frustum'];
+                    scratchFrustum.xOffset = frustum.xOffset;
+                    scratchFrustum.yOffset = frustum.yOffset;
+                    scratchFrustum.fov = frustum.fov;
+                    scratchFrustum.aspectRatio = frustum.aspectRatio;
+                    s.projectionMatrix = Matrix4.clone(scratchFrustum.projectionMatrix, s.projectionMatrix);
+                }
+            }
+            if (!state.entities![EYE_ENTITY_ID] && state['view'] && state['view'].pose) {
+                state.entities![EYE_ENTITY_ID] = state['view'].pose;
+            }
+            // end backwards-compat
+            this._update(state);
+        }
+        this.sessionService.manager.on['ar.context.entityStateMap'] = (entityStateMap: SerializedEntityStateMap) => {
+            for (const id in entityStateMap) {
+                this.updateEntityFromSerializedPose(id, entityStateMap[id]);
+            }
+        }
+
+        scratchFrustum.fov = Math.PI/3;
+        scratchFrustum.aspectRatio = 1;
+
+        this._serializedFrameState = {
+            reality: undefined,
+            time: JulianDate.now(),
+            entities: {},
+            viewport: {x:0,y:0,width:0,height:0},
+            subviews:  [{
+                type: SubviewType.SINGULAR, 
+                viewport: {x:0,y:0,width:1,height:1},
+                projectionMatrix: scratchFrustum.projectionMatrix
+            }],
+        };
+
+        this._update(this._serializedFrameState);
+    }
+
     /**
-     * An event that is raised when all remotely managed entities are are up-to-date for 
-     * the current frame. It is suggested that all modifications to locally managed entities 
-     * should occur within this event. 
+     * An event that is raised when the next frame state is available.
      */
     public frameStateEvent = new Event<FrameState>();
 
     /**
-     * An event that is raised when all remotely managed entities are are up-to-date for 
-     * the current frame. It is suggested that all modifications to locally managed entities 
-     * should occur within this event. 
+     * An event that is raised after managed entities have been updated for 
+     * the current frame. 
      */
     public updateEvent = new Event<ContextService>();
 
@@ -118,26 +187,14 @@ export class ContextService {
     public time = new JulianDate(0,0);
 
     /**
-     * The current cartographic position (WSG84)
-     */
-    public location?:Cartographic;
-    public locationAccuracy?:number;
-    public locationAltitudeAccuracy?:number;
-
-    /**
      * The collection of all entities this application is aware of.
      */
     public entities = new EntityCollection();
 
     /**
-     * An entity representing the location and orientation of the user. 
+     * An alias for the 'eye' entity. To be deprecated in favor of `ViewService.eye`.
      */
-    public user: Entity = this.entities.add(new Entity({
-        id: 'ar.user',
-        name: 'user',
-        position: new ConstantPositionProperty(undefined, ReferenceFrame.FIXED),
-        orientation: new ConstantProperty(Quaternion.IDENTITY)
-    }));
+    public get user() { return this.entities.getById(EYE_ENTITY_ID)! }
 
     /**
      * An entity positioned near the user, aligned with the local East-North-Up
@@ -176,90 +233,81 @@ export class ContextService {
     }
 
     // the current serialized frame state
-    private _serializedFrameState: FrameState = <any>{
-        index: -1,
-        view: undefined
-    };
+    private _serializedFrameState: FrameState;
 
-    private _entityPoseCache: SerializedEntityPoseMap = {};
     private _entityPoseMap = new Map<string, EntityPose | undefined>();
-
-    private _subscribedEntities = new WeakMap<SessionPort, Set<string> | undefined>();
     private _updatingEntities = new Set<string>();
     private _knownEntities = new Set<string>();
-
-    constructor(
-        private sessionService: SessionService) {
-
-        if (this.sessionService.isRealityManager || this.sessionService.isRealityViewer) {
-            this.sessionService.connectEvent.addEventListener((session) => {
-
-                this._subscribedEntities.set(session, new Set<string>());
-
-                session.on['ar.context.subscribe'] = ({id}: { id: string }) => {
-                    const subscriptions = this._subscribedEntities.get(session);
-                    if (subscriptions) subscriptions.add(id);
-                }
-
-            })
-        } else {
-            this.sessionService.manager.on['ar.context.update'] = (state: FrameState) => {
-                this._update(state);
-            }
-        }
-    }
 
     /**
      * Deprecated. Use timestamp property.
      * @private
      */
+    @deprecated('timestamp')
     public get systemTime() {
-        console.warn('systemTime property is deprecated. Use timestamp property');
-        Object.defineProperty(this, 'systemTime',{
-            get: () => this.timestamp
-        });
-        return this.systemTime;
+        return this.timestamp;
     } 
 
     /**
      * Deprecated. To be removed. 
      * @private
      */
-    private getTime(): JulianDate {
-        console.warn('getTime function is deprecated. Use the ContextService.time property instead.')
-        this.getTime = () => this.time;
-        return this.getTime();
+    @deprecated('time')
+    public getTime(): JulianDate {
+        return this.time;
     }
 
     /**
-     * Deprecated. To be removed. 
+     * Deprecated. To be removed. Use the defaultReferenceFrame property. 
      * @private
      */
+    @deprecated('defaultReferenceFrame')
     public setDefaultReferenceFrame(origin: Entity) {
-        console.warn('setDefaultReferenceFrame is deprecated. Use the ContextService.defaultReferenceFrame property directly.')
-        this.setDefaultReferenceFrame = (origin:Entity) => this.defaultReferenceFrame = origin;
-        this.setDefaultReferenceFrame(origin);
+        this.defaultReferenceFrame = origin;
     }
 
     /**
-     * Deprecated. To be removed. 
+     * Deprecated. To be removed.  Use the defaultReferenceFrame property. 
      * @private
      */
-    public getDefaultReferenceFrame(): Entity {        
-        console.warn('getDefaultReferenceFrame is deprecated. Use the ContextService.defaultReferenceFrame property directly.')
-        this.getDefaultReferenceFrame = () => this.defaultReferenceFrame;
-        return this.getDefaultReferenceFrame();
+    @deprecated('defaultReferenceFrame')
+    public getDefaultReferenceFrame(): Entity {
+        return this.defaultReferenceFrame;
     }
 
     /**
-     * Adds an entity to this session's set of tracked entities.
-     *
-     * @param id - The unique identifier of an entity.
-     * @returns The entity that was subscribed to.
+     * Subscribe to pose updates for an entity specified by the given id
+     * 
+     * @deprecated Use [[ContextService#subscribe]]
+     * @param id - the id of the desired entity
+     * @returns A new or existing entity instance matching the given id
      */
+    @deprecated('subscribe')
     public subscribeToEntityById(id: string): Entity {
-        this.sessionService.manager.send('ar.context.subscribe', { id })
+        this.sessionService.manager.send('ar.context.subscribe', { id });
         return this.entities.getOrCreateEntity(id);
+    }
+
+    /**
+     * Subscribe to pose updates for the given entity id
+     * 
+     * @param id - the id of the desired entity
+     * @returns A Promise that resolves to a new or existing entity 
+     * instance matching the given id, if the subscription is successful
+     */
+    public subscribe(id: string|Entity) : Promise<Entity> {
+        id = (<Entity>id).id || id;
+        return this.sessionService.manager.request('ar.context.subscribe', {id}).then(()=>{
+            return this.entities.getOrCreateEntity(id);
+        });
+    }
+
+    /**
+     * Unsubscribe to pose updates for the given entity id
+     */
+    public unsubscribe(id: string|Entity) : void {
+        id = (<Entity>id).id || id;
+        this.sessionService.manager.send('ar.context.unsubscribe', {id});
     }
 
     /**
@@ -271,13 +319,25 @@ export class ContextService {
      * object with the fields `position` and `orientation`, both of type
      * `Cartesian3`. Otherwise undefined.
      */
-    public getEntityPose(entity: Entity, referenceFrame: ReferenceFrame | Entity = this.defaultReferenceFrame): EntityPose {
+    public getEntityPose(entityOrId: Entity|string, referenceFrameOrId: string | ReferenceFrame | Entity = this.defaultReferenceFrame): EntityPose {
         const time = this.time;
+
+        let entity = entityOrId;
+        if (typeof entity === 'string') 
+            entity = this.entities.getById(entityOrId)!;
+        if (!entity) 
+            throw new Error('Unknown entity ' + entityOrId)
+   
+        let referenceFrame = referenceFrameOrId;
+        if (typeof referenceFrame === 'string') 
+            referenceFrame = this.entities.getById(referenceFrameOrId)!;
+        if (!defined(referenceFrame))
+            throw new Error('Unknown entity ' + referenceFrameOrId)
 
         const key = entity.id + '@' + _stringFromReferenceFrame(referenceFrame);
         let entityPose = this._entityPoseMap.get(key);
 
-        if (!defined(entityPose)) {
+        if (!entityPose) {
             entityPose = {
                 position: new Cartesian3,
                 orientation: new Quaternion,
@@ -320,32 +380,59 @@ export class ContextService {
         return entityPose;
     }
 
-    /**
-     * Process the next view state (which should come from the current reality viewer)
-     */
-    public processViewState(view: ViewState) {
-        const frameState = this._serializedFrameState;
-        frameState.index++;
-        frameState.view = view;
-        this._entityPoseCache = {};
-        for (const session of this.sessionService.managedSessions) {
-            if (Role.isRealityAugmenter(session.info.role))
-                this._sendUpdateForSession(frameState, session);
+    public _scratchFrameState:FrameState = {
+        time:<any>{},
+        entities: {},
+        viewport: <any>{},
+        subviews: []
+    }
+
+    public createFrameState(
+        time:JulianDate,
+        viewport:Viewport,
+        subviewList:SerializedSubviewList,
+        eye:Entity,
+        horizontalAccuracy?:number,
+        verticalAccuracy?:number,
+        headingAccuracy?:number
+    ) : FrameState {
+        const eyeMeta = eye['meta'] = eye['meta'] || {};
+        eyeMeta.horizontalAccuracy = horizontalAccuracy || eyeMeta.horizontalAccuracy;
+        eyeMeta.verticalAccuracy = verticalAccuracy || eyeMeta.verticalAccuracy;
+        eyeMeta.headingAccuracy = headingAccuracy || eyeMeta.headingAccuracy;
+
+        for (const s of subviewList) {
+            if (!isFinite(s.projectionMatrix[0]))
+                throw new Error('Invalid projection matrix (contains non-finite values)');
         }
+
+        const frameState:FrameState = this._scratchFrameState;
+        frameState.time = JulianDate.clone(time, frameState.time);
+        frameState.viewport = Viewport.clone(viewport, frameState.viewport);
+        frameState.subviews = SerializedSubviewList.clone(subviewList, frameState.subviews);
+        frameState.entities![EYE_ENTITY_ID] = getSerializedEntityState(eye, time);
+
+        return frameState;
+    }
+
+    private _frameIndex = -1;
+
+    /**
+     * Process the next frame state (which should come from the current reality viewer)
+     */
+    public submitFrameState(frameState: FrameState) {
+        frameState.index = ++this._frameIndex;
         this._update(frameState);
     }
 
     // TODO: This function is called a lot. Potential for optimization. 
-    protected _update(serializedState: FrameState) {
-        
-        // our user entity is defined by the current view pose (the current reality must provide this)
-        this.updateEntityFromSerializedPose(this.user.id, serializedState.view.pose);
+    private _update(frameState: FrameState) {
 
         // update the entities the manager knows about
         this._knownEntities.clear();
-        if (serializedState.entities) {
-            for (const id in serializedState.entities) {
-                this.updateEntityFromSerializedPose(id, serializedState.entities[id]);
+        if (frameState.entities) {
+            for (const id in frameState.entities) {
+                this.updateEntityFromSerializedPose(id, frameState.entities[id]);
                 this._updatingEntities.add(id);
                 this._knownEntities.add(id);
             }
@@ -364,25 +451,28 @@ export class ContextService {
             }
         }
 
-        // update our local origin
-        this._updateLocalOrigin(serializedState);
-
         // update our time values
         const timestamp = performance.now();
         this.deltaTime = Math.min(timestamp - this.timestamp, this.maxDeltaTime);
         this.timestamp = timestamp;
-        JulianDate.clone(<JulianDate>serializedState.view.time, this.time);
-        this._serializedFrameState = serializedState;
+        JulianDate.clone(<JulianDate>frameState.time, this.time);
+
+        // update our stage & local origin. 
+        // TODO: move both of these into the location service, handle in frameStateEvent?
+        this._updateStage(frameState);
+        this._updateLocalOrigin(frameState);
 
         // raise a frame state event (primarily for other services to hook into)
-        this.frameStateEvent.raiseEvent(this._serializedFrameState);
-        // raise eventß for the user to update and render the scene
+        this._serializedFrameState = frameState;
+        this.frameStateEvent.raiseEvent(frameState);
+
+        // raise events for the user to update and render the scene
         this.updateEvent.raiseEvent(this);
         this.renderEvent.raiseEvent(this);
         this.postRenderEvent.raiseEvent(this);
     }
 
-    public updateEntityFromSerializedPose(id:string, entityPose?:SerializedEntityPose) {
+    public updateEntityFromSerializedPose(id:string, entityPose?:SerializedEntityState) {
         const entity = this.entities.getOrCreateEntity(id);
         
         if (!entityPose) {
@@ -416,50 +506,210 @@ export class ContextService {
         return entity;
     }
 
-    public publishEntityState(entity: Entity, referenceFrame: ReferenceFrame|Entity) {
+    private _updateStage(state:FrameState) {
+        // update the stage entity based on the eye entity (provided by the current reality viewer)
+        // and the relative position between the physical eye and the physical stage.
 
-    }
+        const eye = this.entities.getById(EYE_ENTITY_ID);
+        const stage = this.entities.getById(STAGE_ENTITY_ID);
+        const physicalEye = this.entities.getById(PHYSICAL_EYE_ENTITY_ID);
+        const physicalStage = this.entities.getById(PHYSICAL_STAGE_ENTITY_ID);
 
-    private _updateLocalOrigin(state: FrameState) {
-        const pose = state.view.pose;
-        const rootFrame:Entity|ReferenceFrame = 
-            pose ? 
-                typeof pose.r === 'number' ?
-                    pose.r : 
-                    this.entities.getOrCreateEntity(pose.r) : 
-                this.user;
-        const userPosition = this.user.position &&
-            this.user.position.getValueInReferenceFrame(<JulianDate>state.view.time, rootFrame, scratchCartesian3);
-        const localENUFrame = this.localOriginEastNorthUp.position &&
-            this.localOriginEastNorthUp.position.referenceFrame;
-        const localENUPosition = this.localOriginEastNorthUp.position && defined(localENUFrame) &&
-            this.localOriginEastNorthUp.position.getValueInReferenceFrame(<JulianDate>state.view.time, localENUFrame, scratchOriginCartesian3);
-        if (!userPosition || !localENUPosition ||
-            localENUFrame !== rootFrame ||
-            Cartesian3.magnitudeSquared(
-                Cartesian3.subtract(userPosition, localENUPosition, scratchOriginCartesian3)
-            ) > 25000000) {
-                const localENUPositionProperty = <ConstantPositionProperty>this.localOriginEastNorthUp.position;
-                const localENUOrientationProperty = <ConstantProperty>this.localOriginEastNorthUp.orientation;
-            if (userPosition) {
-                localENUPositionProperty.setValue(userPosition, rootFrame);
-                if (rootFrame === ReferenceFrame.FIXED) {
-                    const enuOrientation = Transforms.headingPitchRollQuaternion(userPosition, 0, 0, 0, undefined, scratchQuaternion);
-                    localENUOrientationProperty.setValue(enuOrientation);
-                } else {
-                    localENUOrientationProperty.setValue(Quaternion.IDENTITY);
+        if (!eye || !stage) return;
+
+        stage.position && (stage.position as ConstantPositionProperty).setValue(undefined, undefined);
+        stage.orientation && (stage.orientation as ConstantProperty).setValue(undefined);
+
+        const time = state.time;
+
+        if (physicalEye && physicalStage) {
+            var physicalEyeStageOffset = getEntityPositionInReferenceFrame(
+                physicalEye, 
+                time, 
+                physicalStage, 
+                scratchCartesian
+            );
+        }
+
+        if (!physicalEyeStageOffset) {
+            physicalEyeStageOffset = Cartesian3.fromElements(0,0, AVERAGE_HUMAN_HEIGHT, scratchCartesian);
+        }
+
+        const eyePositionFixed = getEntityPositionInReferenceFrame(
+            eye,
+            time,
+            ReferenceFrame.FIXED,
+            scratchCartesian2
+        );
+        
+        if (eyePositionFixed) {
+            const enuToFixedFrameTransform = Transforms.eastNorthUpToFixedFrame(eyePositionFixed, undefined, scratchMatrix4);
+            const enuRotationMatrix = Matrix4.getRotation(enuToFixedFrameTransform, scratchMatrix3);
+            const enuOrientation = Quaternion.fromRotationMatrix(enuRotationMatrix);
+            const physicalEyeStageOffsetFixed = Matrix3.multiplyByVector(enuRotationMatrix, physicalEyeStageOffset, physicalEyeStageOffset);
+            const stagePositionFixed = Cartesian3.subtract(eyePositionFixed, physicalEyeStageOffsetFixed, physicalEyeStageOffsetFixed);
+            stage.position = stage.position || new ConstantPositionProperty();
+            stage.orientation = stage.orientation || new ConstantProperty();
+            (stage.position as ConstantPositionProperty).setValue(stagePositionFixed, ReferenceFrame.FIXED);
+            (stage.orientation as ConstantProperty).setValue(enuOrientation);
+        } else {        
+            const eyeFrame = eye && eye.position ? eye.position.referenceFrame : undefined;
+            if (eyeFrame) {
+                const eyePositionRelativeToEyeFrame = getEntityPositionInReferenceFrame(
+                    eye,
+                    time,
+                    eyeFrame,
+                    scratchCartesian2
+                );
+                if (eyePositionRelativeToEyeFrame) {
+                    const stagePositionRelativeToEye = Cartesian3.subtract(eyePositionRelativeToEyeFrame, physicalEyeStageOffset, physicalEyeStageOffset);
+                    (stage.position as ConstantPositionProperty).setValue(stagePositionRelativeToEye, eyeFrame);
+                    (stage.orientation as ConstantProperty).setValue(Quaternion.IDENTITY);
                 }
-            } else {
-                localENUPositionProperty.setValue(Cartesian3.ZERO, this.user);
-                localENUOrientationProperty.setValue(Quaternion.IDENTITY);
             }
-            this.localOriginChangeEvent.raiseEvent(undefined);
         }
     }
 
-    private _sessionEntities:SerializedEntityPoseMap = {};
+    private _updateLocalOrigin(state: FrameState) {
+        const eye = this.entities.getById(EYE_ENTITY_ID);
+        const stage = this.entities.getById(STAGE_ENTITY_ID);
 
-    private _sendUpdateForSession(state: FrameState, session: SessionPort) {
+        const stageFrame = stage && stage.position ? stage.position.referenceFrame : undefined;
+        if (!eye || !stage) return;
+
+        if (!defined(stageFrame)) {
+            if (this.localOriginEastNorthUp.position!.referenceFrame !== stage) {
+                (this.localOriginEastNorthUp.position as ConstantPositionProperty).setValue(Cartesian3.ZERO, stage);
+                (this.localOriginEastNorthUp.orientation as ConstantProperty).setValue(Quaternion.IDENTITY);
+                this.localOriginChangeEvent.raiseEvent(undefined);
+            }
+            return;
+        }
+        
+        const eyePosition =
+            eye.position && eye.position.getValueInReferenceFrame(state.time, stageFrame, scratchCartesian);
+
+        if (!eyePosition) return;
+        
+        const localOriginPosition = this.localOriginEastNorthUp
+            .position!.getValueInReferenceFrame(state.time, stageFrame, scratchOriginCartesian);
+        
+        if (!localOriginPosition ||
+            Cartesian3.magnitude(
+                Cartesian3.subtract(eyePosition, localOriginPosition, scratchOriginCartesian)
+            ) > 5000) {
+                const localOriginPositionProperty = <ConstantPositionProperty>this.localOriginEastNorthUp.position;
+                const localOriginOrientationProperty = <ConstantProperty>this.localOriginEastNorthUp.orientation;
+                const stagePosition = stage.position && stage.position.getValueInReferenceFrame(state.time, stageFrame, scratchCartesian);
+                const stageOrientation = stage.orientation && stage.orientation.getValue(state.time, scratchQuaternion);
+                localOriginPositionProperty.setValue(stagePosition, stageFrame)
+                localOriginOrientationProperty.setValue(stageOrientation);
+                this.localOriginChangeEvent.raiseEvent(undefined);
+        }
+    }
+}
+
+function _stringFromReferenceFrame(referenceFrame: string | ReferenceFrame | Entity): string {
+    const rf = referenceFrame as Entity;
+    return defined(rf.id) ? rf.id : '' + rf;
+}
+
+
+
+@autoinject()
+export class ContextServiceProvider {
+
+    public entitySubscriptionsBySubscriber = new WeakMap<SessionPort, Set<string>>();
+    public subscribersByEntityId = new Map<string, Set<SessionPort>>();
+    public subscribersChangeEvent = new Event<{id:string, subscribers}>();
+
+    public publishingReferenceFrameMap = new Map<string, string|ReferenceFrame>();
+    
+    private _entityPoseCache: SerializedEntityStateMap = {};
+
+    constructor(
+        private sessionService:SessionService,
+        private contextService:ContextService
+    ) {
+        sessionService.connectEvent.addEventListener((session) => {
+            const  subscriptions = new Set<string>();
+            this.entitySubscriptionsBySubscriber.set(session, subscriptions);
+
+            session.on['ar.context.subscribe'] = ({id}:{id:string}) => {
+                const subscribers = this.subscribersByEntityId.get(id) || new Set<SessionPort>();
+                this.subscribersByEntityId.set(id, subscribers);
+                subscribers.add(session);
+                subscriptions.add(id);
+                this.subscribersChangeEvent.raiseEvent({id, subscribers});
+
+                session.closeEvent.addEventListener(()=>{
+                    subscribers.delete(session);
+                    this.subscribersChangeEvent.raiseEvent({id, subscribers});
+                })
+            }
+
+            session.on['ar.context.unsubscribe'] = ({id}:{id:string}) => {
+                const subscribers = this.subscribersByEntityId.get(id);
+                subscribers && subscribers.delete(session);
+                subscriptions.delete(id);
+                this.subscribersChangeEvent.raiseEvent({id, subscribers});
+            }
+
+            session.closeEvent.addEventListener(()=>{
+                subscriptions.forEach((id)=>{
+                    const subscribers = this.subscribersByEntityId.get(id);
+                    subscribers && subscribers.delete(session);
+                    this.subscribersChangeEvent.raiseEvent({id, subscribers});
+                });
+                this.entitySubscriptionsBySubscriber.delete(session);
+            })
+        });
+
+        this.contextService.updateEvent.addEventListener(()=>{
+            this._publishUpdates();
+        });
+    }
+
+    public publishEntityState(idOrEntity:string|Entity, time=this.contextService.time) {
+        let id:string;
+        let entity:Entity|undefined;
+
+        if ((<Entity>idOrEntity).id) {
+            entity = <Entity>idOrEntity;
+            id = entity.id;
+        } else {
+            id = <string>idOrEntity;
+            entity = this.contextService.entities.getById(id);
+        }
+
+        const subscribers = this.subscribersByEntityId.get(id);
+
+        if (entity && subscribers) {
+            const referenceFrameId = this.publishingReferenceFrameMap.get(id);
+            const referenceFrame = defined(referenceFrameId) && typeof referenceFrameId === 'string' ? 
+                this.contextService.entities.getById(referenceFrameId) :
+                referenceFrameId;
+
+            const entityStateMap = {[id]: getSerializedEntityState(entity, time, referenceFrame)};
+            
+            for (const s of subscribers) {
+                s.send('ar.context.entityStateMap', entityStateMap);
+            }
+        }
+    }
+    
+    private _publishUpdates() {
+        this._entityPoseCache = {};
+        for (const session of this.sessionService.managedSessions) {
+            if (Role.isRealityAugmenter(session.info.role))
+                this._sendUpdateForSession(session);
+        }
+    }
+
+    private _sessionEntities:SerializedEntityStateMap = {};
+
+    private _sendUpdateForSession(session: SessionPort) {
+        const state = this.contextService.serializedFrameState!;
         const sessionEntities = this._sessionEntities;
 
         // clear session entities
@@ -475,30 +725,33 @@ export class ContextService {
         }
 
         // get subscrbied entities for the session
-        for (const id of <string[]><any>this._subscribedEntities.get(session)) {
-            sessionEntities[id] = this._getSerializedEntityPose(id, <JulianDate>state.view.time);
+        for (const id of <string[]><any>this.entitySubscriptionsBySubscriber.get(session)) {
+            const entity = this.contextService.entities.getById(id);
+            sessionEntities[id] = this._getSerializedEntityState(entity, <JulianDate>state.time);
         }
 
         // recycle the parent frame state object, but with the session entities
         const parentEntities = state.entities;
         state.entities = sessionEntities;
-        state['time'] = state.view.time; // deprecated
+        state.time = state.time;
         state.sendTime = JulianDate.now(<JulianDate>state.sendTime);
         if (session.info.version) session.send('ar.context.update', state);
         state.entities = parentEntities;
     }
 
-    private _getSerializedEntityPose(id: string, time: JulianDate) {
+    private _getSerializedEntityState(entity: Entity|undefined, time: JulianDate) {
+        if (!entity) return undefined;
+
+        const id = entity.id;
+
         if (!defined(this._entityPoseCache[id])) {
-            const entity = this.entities.getById(id);
-            if (!entity) return undefined;
-            this._entityPoseCache[id] = getSerializedEntityPose(entity, time);
+            const referenceFrameId = this.publishingReferenceFrameMap.get(id);
+            const referenceFrame = defined(referenceFrameId) && typeof referenceFrameId === 'string' ? 
+                this.contextService.entities.getById(referenceFrameId) :
+                referenceFrameId;
+            this._entityPoseCache[id] = getSerializedEntityState(entity, time, referenceFrame);
         }
+
         return this._entityPoseCache[id];
     }
-}
-
-function _stringFromReferenceFrame(referenceFrame: ReferenceFrame | Entity): string {
-    const rf = referenceFrame as Entity;
-    return defined(rf.id) ? rf.id : '' + rf;
 }
