@@ -1,10 +1,9 @@
 import { SessionService, SessionPort } from './session';
 import { Event } from './utils';
 /**
- * Manages focus state
+ * Access focus state
  */
 export declare class FocusService {
-    private sessionService;
     /**
      * An event that is raised when this app has gained focus
      */
@@ -18,23 +17,18 @@ export declare class FocusService {
      */
     readonly hasFocus: boolean;
     private _hasFocus;
-    /**
-     * Manager-only. An event that is raised when a managed session has acquired focus.
-     */
-    readonly sessionFocusEvent: Event<{
+    constructor(sessionService: SessionService);
+}
+/**
+ * Manage focus state
+ */
+export declare class FocusServiceProvider {
+    private sessionService;
+    sessionFocusEvent: Event<{
         previous?: SessionPort | undefined;
         current?: SessionPort | undefined;
     }>;
-    private _sessionFocusEvent;
-    private _session?;
     constructor(sessionService: SessionService);
-    /**
-     * Manager-only. The managed session which currently has focus.
-     */
-    /**
-     *  Manager-only. Grant focus to a managed session.
-     */
+    private _session?;
     session: SessionPort | undefined;
-    whenSessionHasFocus(session: SessionPort): Promise<{}>;
-    private _setFocus(state);
 }
