@@ -3,7 +3,7 @@ import { SessionService, SessionPort } from './session';
 import { Event } from './utils';
 
 /**
- * Manages focus state
+ * Access focus state
  */
 @inject(SessionService)
 export class FocusService {
@@ -24,45 +24,43 @@ export class FocusService {
     public get hasFocus() { return this._hasFocus }
     private _hasFocus = false;
 
-    /**
-     * Manager-only. An event that is raised when a managed session has acquired focus. 
-     */
-    public get sessionFocusEvent() {
-        this.sessionService.ensureIsRealityManager();
-        return this._sessionFocusEvent;
-    }
-    private _sessionFocusEvent = new Event<{ previous?: SessionPort, current?: SessionPort }>();
-
-    private _session?: SessionPort;
-
-    constructor(private sessionService: SessionService) {
-        sessionService.manager.on['ar.focus.state'] = (message: { state: boolean }) => {
-            this._setFocus(message.state);
+    constructor(sessionService: SessionService) {
+        sessionService.manager.on['ar.focus.state'] = ({state}: { state: boolean }) => {
+            if (this._hasFocus !== state) {
+                this._hasFocus = state;
+                if (state) {
+                    this.focusEvent.raiseEvent(undefined);
+                } else {
+                    this.blurEvent.raiseEvent(undefined);
+                }
+            }
         }
+    }
+}
 
-        if (sessionService.isRealityManager) {
-            sessionService.manager.connectEvent.addEventListener(() => {
-                setTimeout(() => {
-                    if (!this._session && this.sessionService.manager.isConnected)
-                        this.session = this.sessionService.manager;
-                })
+
+/**
+ * Manage focus state
+ */
+@inject(SessionService, FocusService)
+export class FocusServiceProvider {
+
+    public sessionFocusEvent = new Event<{ previous?: SessionPort, current?: SessionPort }>();
+
+    constructor(private sessionService:SessionService) {
+        sessionService.ensureIsRealityManager();
+        sessionService.manager.connectEvent.addEventListener(() => {
+            setTimeout(() => {
+                if (!this._session && this.sessionService.manager.isConnected)
+                    this.session = this.sessionService.manager;
             })
-        }
+        });
     }
 
-    /**
-     * Manager-only. The managed session which currently has focus. 
-     */
-    public get session() {
-        this.sessionService.ensureIsRealityManager();
-        return this._session;
-    }
+    private _session?:SessionPort;
 
-    /**
-     *  Manager-only. Grant focus to a managed session.
-     */
+    public get session() { return this._session };
     public set session(session: SessionPort|undefined) {
-        this.sessionService.ensureIsRealityManager();
         if (session && !session.isConnected)
             throw new Error('Only a connected session can be granted focus')
         const previousFocussedSession = this._session;
@@ -77,28 +75,4 @@ export class FocusService {
             });
         }
     }
-
-    public whenSessionHasFocus(session: SessionPort) {
-        this.sessionService.ensureIsRealityManager();
-        return new Promise((resolve) => {
-            let remove = this.sessionFocusEvent.addEventListener(({current}) => {
-                if (current === session) {
-                    remove();
-                    resolve();
-                }
-            });
-        });
-    }
-
-    private _setFocus(state) {
-        if (this._hasFocus !== state) {
-            this._hasFocus = state;
-            if (state) {
-                this.focusEvent.raiseEvent(undefined);
-            } else {
-                this.blurEvent.raiseEvent(undefined);
-            }
-        }
-    }
-
 }
