@@ -1,25 +1,42 @@
-var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
-    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
-    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
-    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
-    return c > 3 && r && Object.defineProperty(target, key, r), r;
-};
-import { inject } from 'aurelia-dependency-injection';
-import { SessionService } from '../session';
+import { Event } from '../utils';
 /**
  * Abstract class for a RealityViewer
  */
-export let RealityViewer = class RealityViewer {
-    constructor(_sessionService, uri) {
-        this._sessionService = _sessionService;
+export class RealityViewer {
+    constructor(uri) {
         this.uri = uri;
+        this.providedReferenceFrames = [];
+        this.connectEvent = new Event();
+        this.presentChangeEvent = new Event();
+        this._isPresenting = false;
+        this.connectEvent.addEventListener((session) => {
+            if (this._session)
+                this._session.close();
+            this._session = session;
+            session.closeEvent.addEventListener(() => {
+                if (this._session === session)
+                    this._session = undefined;
+            });
+        });
     }
-    load() {
-        if (this.session)
+    get isPresenting() {
+        return this._isPresenting;
+    }
+    get session() {
+        return this._session;
+    }
+    destroy() {
+        if (this.session) {
             this.session.close();
-        this.session = this._sessionService.addManagedSessionPort(this.uri);
+        }
     }
     ;
+    setPresenting(flag) {
+        if (this._isPresenting !== flag) {
+            this._isPresenting = flag;
+            this.presentChangeEvent.raiseEvent(undefined);
+        }
+    }
     static getType(uri) {
         if (uri === undefined)
             return undefined;
@@ -28,9 +45,7 @@ export let RealityViewer = class RealityViewer {
         }
         return 'hosted';
     }
-};
+}
+RealityViewer.DEFAULT = 'reality:default';
 RealityViewer.EMPTY = 'reality:empty';
 RealityViewer.LIVE = 'reality:live';
-RealityViewer = __decorate([
-    inject(SessionService)
-], RealityViewer);
