@@ -1,22 +1,22 @@
-import { SessionService } from '../session';
-import { DeviceService } from '../device';
+import { SessionService, SessionPort } from '../session';
 import { ViewService } from '../view';
-import { VuforiaServiceDelegate } from '../vuforia';
+import { ViewportService } from '../viewport';
+import { ContextService } from '../context';
 import { RealityViewer } from './base';
 export declare class LiveRealityViewer extends RealityViewer {
     private sessionService;
-    private vuforiaDelegate;
+    private viewportService;
     private viewService;
-    private deviceService;
+    private contextService;
     uri: string;
     videoElement: HTMLVideoElement;
     private canvas;
     private context;
     private videoFov;
     private settingsIframe;
-    constructor(sessionService: SessionService, vuforiaDelegate: VuforiaServiceDelegate, viewService: ViewService, deviceService: DeviceService, uri: string);
+    constructor(sessionService: SessionService, viewportService: ViewportService, viewService: ViewService, contextService: ContextService, uri: string);
     destroy(): void;
-    setPresenting(foo: boolean): void;
+    protected setupInternalSession(session: SessionPort): void;
     load(): void;
     static isAvailable(): boolean;
     getVideoFrame(x: number, y: number, width: number, height: number): ImageData;
