@@ -4,21 +4,23 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
+var __metadata = (this && this.__metadata) || function (k, v) {
+    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+};
 import { inject } from 'aurelia-dependency-injection';
 import { Role } from '../common';
 import { SessionService } from '../session';
-import { DeviceService } from '../device';
 import { ViewService } from '../view';
-import { VuforiaServiceDelegate } from '../vuforia';
-import * as utils from '../utils';
+import { ViewportService } from '../viewport';
+import { ContextService } from '../context';
 import { RealityViewer } from './base';
-export let LiveRealityViewer = class LiveRealityViewer extends RealityViewer {
-    constructor(sessionService, vuforiaDelegate, viewService, deviceService, uri) {
-        super(sessionService, uri);
+let LiveRealityViewer = class LiveRealityViewer extends RealityViewer {
+    constructor(sessionService, viewportService, viewService, contextService, uri) {
+        super(uri);
         this.sessionService = sessionService;
-        this.vuforiaDelegate = vuforiaDelegate;
+        this.viewportService = viewportService;
         this.viewService = viewService;
-        this.deviceService = deviceService;
+        this.contextService = contextService;
         this.uri = uri;
         if (typeof document !== 'undefined') {
             this.settingsIframe = document.createElement('iframe');
@@ -33,97 +35,97 @@ export let LiveRealityViewer = class LiveRealityViewer extends RealityViewer {
             this.videoElement.controls = false;
             this.videoElement.autoplay = true;
             this.videoElement.style.display = 'none';
-            const viewElement = this.viewService.element;
+            const viewElement = this.viewportService.element;
             viewElement.insertBefore(this.settingsIframe, viewElement.firstChild);
             viewElement.insertBefore(this.videoElement, viewElement.firstChild);
             this.canvas = document.createElement('canvas');
             this.context = this.canvas.getContext('2d');
-        }
-    }
-    destroy() {
-        this.settingsIframe.remove();
-        this.videoElement.remove();
-        this.canvas.remove();
-    }
-    setPresenting(foo) {
-        this.videoElement.style.display = foo ? 'initial' : 'none';
-    }
-    load() {
-        super.load();
-        const realitySession = this.session;
-        const remoteRealitySession = this.sessionService.createSessionPort(this.uri);
-        remoteRealitySession.on['ar.device.state'] = () => { };
-        remoteRealitySession.on['ar.view.uievent'] = () => { };
-        remoteRealitySession.on['ar.context.update'] = () => { };
-        remoteRealitySession.on['ar.reality.connect'] = () => { };
-        remoteRealitySession.connectEvent.addEventListener(() => {
-            if (this.videoElement) {
-                const videoElement = this.videoElement;
-                const mediaDevices = navigator.mediaDevices;
-                const getUserMedia = (mediaDevices.getUserMedia || mediaDevices['mozGetUserMedia'] ||
-                    mediaDevices['msGetUserMedia'] || mediaDevices['webkitGetUserMedia']).bind(mediaDevices);
-                getUserMedia({ audio: false, video: true }).then((videoStream) => {
-                    videoElement.src = window.URL.createObjectURL(videoStream);
-                }).catch((error) => {
-                    remoteRealitySession.errorEvent.raiseEvent(error);
-                });
-                const deviceService = this.deviceService;
-                let lastFrameTime = -1;
-                let update = (time) => {
-                    if (realitySession.isConnected)
-                        this.deviceService.requestFrame(update);
-                    else
-                        return;
-                    if (videoElement.currentTime != lastFrameTime) {
-                        lastFrameTime = videoElement.currentTime;
-                        // const videoWidth = videoElement.videoWidth;
-                        // const videoHeight = videoElement.videoHeight;
-                        const viewState = {
-                            time,
-                            pose: utils.getSerializedEntityPose(deviceService.eye, time),
-                            geolocationAccuracy: deviceService.geolocationAccuracy,
-                            altitudeAccuracy: deviceService.altitudeAccuracy,
-                            compassAccuracy: deviceService.compassAccuracy,
-                            viewport: deviceService.viewport,
-                            subviews: deviceService.subviews
-                        };
-                        remoteRealitySession.send('ar.reality.viewState', viewState);
-                    }
-                };
-                this.deviceService.requestFrame(update);
-            }
-            this.vuforiaDelegate.videoEnabled = true;
-            this.vuforiaDelegate.trackingEnabled = true;
-            const remove = this.vuforiaDelegate.stateUpdateEvent.addEventListener((viewState) => {
-                remoteRealitySession.send('ar.reality.viewState', viewState);
-            });
-            remoteRealitySession.closeEvent.addEventListener(() => {
-                remove();
-                this.vuforiaDelegate.videoEnabled = false;
-                this.vuforiaDelegate.trackingEnabled = false;
-            });
-        });
-        if (typeof document !== 'undefined' && typeof navigator !== 'undefined') {
             window.addEventListener('message', (event) => {
                 const origin = event.origin;
                 if (origin === 'http://argonjs.io') {
                     this.videoFov = event.data; // TODO: this is not flexible. Should be passing an object with message type and data
                 }
             });
-            const mediaDevices = navigator.mediaDevices;
-            const getUserMedia = (mediaDevices.getUserMedia || mediaDevices['mozGetUserMedia'] ||
-                mediaDevices['msGetUserMedia'] || mediaDevices['webkitGetUserMedia']).bind(mediaDevices);
-            getUserMedia({ audio: false, video: true }).then((videoStream) => {
-                this.videoElement.src = window.URL.createObjectURL(videoStream);
-            }).catch((error) => {
-                remoteRealitySession.errorEvent.raiseEvent(error);
-            });
         }
+        this.presentChangeEvent.addEventListener(() => {
+            if (typeof document !== 'undefined') {
+                this.videoElement.style.display = this.isPresenting ? 'initial' : 'none';
+            }
+        });
+    }
+    destroy() {
+        super.destroy();
+        if (typeof document !== 'undefined') {
+            this.settingsIframe.remove();
+            this.videoElement.remove();
+            this.canvas.remove();
+        }
+    }
+    setupInternalSession(session) {
+        session.on['ar.device.state'] = () => { };
+        session.on['ar.visibility.state'] = () => { };
+        session.on['ar.focus.state'] = () => { };
+        session.on['ar.viewport.presentationMode'] = () => { };
+        session.on['ar.viewport.uievent'] = () => { };
+        session.on['ar.view.suggestedViewState'] = () => { };
+        session.on['ar.context.update'] = () => { };
+        session.on['ar.reality.connect'] = () => { };
+        session.connectEvent.addEventListener(() => {
+            if (this.videoElement) {
+                const videoElement = this.videoElement;
+                const mediaDevices = navigator.mediaDevices;
+                const getUserMedia = (mediaDevices.getUserMedia || mediaDevices['mozGetUserMedia'] ||
+                    mediaDevices['msGetUserMedia'] || mediaDevices['webkitGetUserMedia']).bind(mediaDevices);
+                getUserMedia({ audio: false, video: true }).then((videoStream) => {
+                    const stopVideoStream = () => {
+                        for (const t of videoStream.getTracks()) {
+                            t.stop();
+                        }
+                    };
+                    if (session.isConnected) {
+                        videoElement.src = window.URL.createObjectURL(videoStream);
+                        session.closeEvent.addEventListener(stopVideoStream);
+                    }
+                    else {
+                        stopVideoStream();
+                    }
+                }).catch((error) => {
+                    session.errorEvent.raiseEvent(error);
+                });
+                const viewService = this.viewService;
+                let lastFrameTime = -1;
+                let update = (time) => {
+                    if (session.isConnected)
+                        viewService.requestAnimationFrame(update);
+                    else
+                        return;
+                    const suggestedViewState = viewService.suggestedViewState;
+                    if (!suggestedViewState)
+                        return;
+                    if (videoElement.currentTime != lastFrameTime) {
+                        lastFrameTime = videoElement.currentTime;
+                        // const videoWidth = videoElement.videoWidth;
+                        // const videoHeight = videoElement.videoHeight;
+                        const frameState = this.contextService.createFrameState(time, suggestedViewState.viewport, suggestedViewState.subviews, viewService.eye);
+                        session.send('ar.reality.frameState', frameState);
+                    }
+                };
+                viewService.requestAnimationFrame(update);
+            }
+        });
+    }
+    load() {
+        const session = this.sessionService.addManagedSessionPort(this.uri);
+        session.connectEvent.addEventListener(() => {
+            this.connectEvent.raiseEvent(session);
+        });
+        const internalSession = this.sessionService.createSessionPort(this.uri);
+        this.setupInternalSession(internalSession);
         // Only connect after the caller is able to attach connectEvent handlers
         Promise.resolve().then(() => {
             const messageChannel = this.sessionService.createSynchronousMessageChannel();
-            realitySession.open(messageChannel.port1, this.sessionService.configuration);
-            remoteRealitySession.open(messageChannel.port2, { role: Role.REALITY_VIEWER, title: 'Local' });
+            session.open(messageChannel.port1, this.sessionService.configuration);
+            internalSession.open(messageChannel.port2, { role: Role.REALITY_VIEWER, title: 'Live' });
         });
     }
     static isAvailable() {
@@ -143,5 +145,10 @@ export let LiveRealityViewer = class LiveRealityViewer extends RealityViewer {
     }
 };
 LiveRealityViewer = __decorate([
-    inject(SessionService, VuforiaServiceDelegate, ViewService, DeviceService)
+    inject(SessionService, ViewportService, ViewService, ContextService),
+    __metadata("design:paramtypes", [SessionService,
+        ViewportService,
+        ViewService,
+        ContextService, String])
 ], LiveRealityViewer);
+export { LiveRealityViewer };
