@@ -1,13 +1,16 @@
-import { ViewService } from './view';
+import { ViewServiceProvider } from './view';
+import { ViewportService } from './viewport';
 import { SessionService } from './session';
-import { RealityService } from './reality';
+import { RealityService, RealityServiceProvider } from './reality';
 /**
  * Provides a default UI
  */
 export declare class DefaultUIService {
     private sessionService;
-    private viewService;
+    private viewportService;
     private realityService;
+    private realityServiceProvider;
+    private viewServiceProvider;
     private element?;
     private realityViewerSelectorElement;
     private realityViewerListElement;
@@ -19,7 +22,7 @@ export declare class DefaultUIService {
     private hmdMenuItem;
     private realityMenuItem;
     private maximizeMenuItem;
-    constructor(sessionService: SessionService, viewService: ViewService, realityService: RealityService);
+    constructor(sessionService: SessionService, viewportService: ViewportService, realityService: RealityService, realityServiceProvider: RealityServiceProvider, viewServiceProvider: ViewServiceProvider);
     private _createMenuItem(icon, hint, onSelect?);
     private onSelect(element, cb);
     private toggleMenu();
