@@ -1,80 +1,18 @@
 /// <reference types="cesium" />
-import { SerializedEntityPose } from './common';
+import { SerializedEntityState } from './common';
 import { Entity, JulianDate, PerspectiveFrustum, PerspectiveOffCenterFrustum, Quaternion, Cartesian3, ReferenceFrame, Matrix4 } from './cesium/cesium-imports';
+export * from './utils/command-queue';
+export * from './utils/event';
+export * from './utils/message-channel';
+export { default as synthesizeEvent } from './utils/ui-event-synthesizer';
+export { default as createEventForwarder } from './utils/ui-event-forwarder';
 /**
- * A callback for removing the event listener.
- */
-export declare type RemoveCallback = () => void;
-/**
- * Provides the ability raise and subscribe to an event.
- */
-export declare class Event<T> {
-    private _event;
-    /**
-     * Get the number of listeners currently subscribed to the event.
-     * @return Number of listeners currently subscribed to the event.
-     */
-    readonly numberOfListeners: number;
-    /**
-      * Add an event listener.
-      * @param The function to be executed when the event is raised.
-      * @return A convenience function which removes this event listener when called
-      */
-    addEventListener(listener: (data: T) => void): RemoveCallback;
-    /**
-     * Remove an event listener.
-     * @param The function to be unregistered.
-     * @return True if the listener was removed;
-     * false if the listener and scope are not registered with the event.
-     */
-    removeEventListener(listener: (data: T) => void): boolean;
-    /**
-     * Raises the event by calling each registered listener with all supplied arguments.
-     * @param This method takes any number of parameters and passes them through to the listener functions.
-     */
-    raiseEvent(data: T): void;
-}
-/**
-* TODO.
-*/
-export declare class CommandQueue {
-    private _queue;
-    private _currentCommand;
-    private _currentCommandPending;
-    private _paused;
-    /**
-     * An error event.
-     */
-    errorEvent: Event<Error>;
-    /**
-     * If errorEvent has 1 listener, outputs the error message to the web console.
-     */
-    constructor();
-    /**
-     * Push a command to the command queue.
-     * @param command Any command ready to be pushed into the command queue.
-     */
-    push<TResult>(command: () => TResult, execute?: boolean): Promise<TResult>;
-    /**
-     * Execute the command queue
-     */
-    execute(): void;
-    /**
-     * Puase the command queue (currently executing commands will still complete)
-     */
-    pause(): void;
-    /**
-     * Clear commandQueue.
-     */
-    clear(): void;
-    private _executeNextCommand();
-}
-/**
- * Get array of ancestor reference frames of a Cesium Entity.
+ * Get array of ancestor reference frames of a Cesium Entity, ordered from
+ * farthest ancestor to the passed frame.
  * @param frame A Cesium Entity to get ancestor reference frames.
  * @param frames An array of reference frames of the Cesium Entity.
  */
-export declare function getAncestorReferenceFrames(frame: Entity): (Entity | ReferenceFrame)[];
+export declare function getAncestorReferenceFrames(frame: Entity, result?: never[]): (Entity | ReferenceFrame)[];
 /**
  * Gets the value of the Position property at the provided time and in the provided reference frame.
  * @param entity The entity to get position.
@@ -110,7 +48,7 @@ export declare const getEntityOrientation: typeof getEntityOrientationInReferenc
  * serialized according to the furthest ancestor frame that resolves to a valid pose.
  * @return An EntityPose object with orientation, position and referenceFrame.
  */
-export declare function getSerializedEntityPose(entity: Entity, time: JulianDate, frame?: ReferenceFrame | Entity): SerializedEntityPose | undefined;
+export declare function getSerializedEntityState(entity: Entity, time: JulianDate, frame?: ReferenceFrame | Entity): SerializedEntityState | undefined;
 /**
  * If urlParser does not have a value, throw error message "resolveURL requires DOM api".
  * If inURL is undefined, throw error message "expected inURL".
@@ -136,89 +74,7 @@ export declare function parseURL(inURL: string): {
     hash: string;
     host: string;
 };
-export declare function resolveElement(elementOrSelector: string | HTMLElement): Promise<HTMLElement>;
-/**
- * Returns a viewport that reflects the size of the current window
- */
-export declare function getWindowViewport(): {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-};
-/**
- * A minimal MessageEvent interface.
- */
-export declare class MessageEventLike {
-    constructor(data: any);
-    data: any;
-}
-/**
- * A minimal MessagePort interface.
- */
-export interface MessagePortLike {
-    /**
-      * A callback for handling incoming messages.
-      */
-    onmessage?: (ev: MessageEventLike) => any;
-    /**
-     * Send a message through this message port.
-     * @param message The message needed to be posted.
-     */
-    postMessage(message?: any): void;
-    /**
-     * Close this message port.
-     */
-    close?: () => void;
-}
-/**
- * A MessageChannel pollyfill.
- */
-export declare class MessageChannelLike {
-    /**
-     * The first port.
-     */
-    port1: MessagePortLike;
-    /**
-     * The second port.
-     */
-    port2: MessagePortLike;
-    /**
-     * Create a MessageChannelLike instance.
-     */
-    constructor();
-}
-/**
- * A synchronous MessageChannel.
- */
-export declare class SynchronousMessageChannel {
-    /**
-     * The first port.
-     */
-    port1: MessagePortLike;
-    /**
-     * The second port.
-     */
-    port2: MessagePortLike;
-    /**
-     * Create a MessageChannelLike instance.
-     */
-    constructor();
-}
-/**
- * A factory which creates MessageChannel or MessageChannelLike instances, depending on
- * wheter or not MessageChannel is avaialble in the execution context.
- */
-export declare class MessageChannelFactory {
-    /**
-     * Create a MessageChannel (or MessageChannelLike) instance.
-     */
-    create(): MessageChannelLike;
-    /**
-     * Create a SynchronousMessageChannel instance.
-     */
-    createSynchronous(): SynchronousMessageChannel;
-}
+export declare function resolveElement(elementOrSelector: string | HTMLElement): Promise<{}>;
 export declare function decomposePerspectiveOffCenterProjectionMatrix(mat: Matrix4, result: PerspectiveOffCenterFrustum): PerspectiveOffCenterFrustum;
 export declare function decomposePerspectiveProjectionMatrix(mat: Matrix4, result: PerspectiveFrustum): PerspectiveFrustum;
 /**
@@ -234,3 +90,4 @@ export declare function convertEntityReferenceFrame(entity: Entity, time: Julian
 export declare const isIOS: boolean;
 export declare function openInArgonApp(): void;
 export declare const requestAnimationFrame: any;
+export declare function deprecated(alternative?: string): MethodDecorator;
