@@ -1,16 +1,5 @@
-declare const global;
-
-if (typeof document == 'undefined') { // check for non-DOM environment
-    const w = typeof global !== 'undefined' ? global : window;
-    // global.window = w;
-    w['WebVRConfig'] = {
-        DEFER_INITIALIZATION: typeof navigator == 'undefined'
-    };
-}
-
 import CardboardUI from 'googlevr/webvr-polyfill/src/cardboard-ui';
-
-console.log(CardboardUI);
+import 'googlevr/webvr-polyfill/src/main';
 
 // fix cardboard ui to listen for touch or mouse events.
 // see https://github.com/googlevr/webvr-polyfill/issues/174
@@ -36,5 +25,3 @@ CardboardUI.prototype.listen = function(optionsCallback, backCallback) {
   canvas.addEventListener('click', this.listener, false);
   canvas.addEventListener('touchstart', this.listener, false); // for some reason, using `touchend` here breaks touch events in parent DOM elements, whereas touchstart works fine.... seems like a bug in WebKit
 };
-
-import 'googlevr/webvr-polyfill/src/main'
