@@ -1,5 +1,5 @@
-declare module '*package.json!' {
-    export const version : number;
+declare module '*package.json' {
+    export const version : string;
 }
 
 declare module 'googlevr/webvr-polyfill/src/cardboard-ui';
